ShopBackend/requirements.txt
altgraph==0.17.4
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==24.4.2
Automat==22.10.0
bcrypt==3.2.0
billiard==4.2.0
blinker==1.4
celery==5.4.0
certifi==2020.6.20
cffi==1.16.0
channels==3.0.5
channels-redis==4.1.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.4
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==43.0.0
daphne==3.0.2
distro==1.7.0
Django==4.2.7
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-comment-migrate==0.1.7
django-cors-headers==4.3.0
django-extensions==3.2.3
django-filter==23.3
django-ranged-response==0.2.0
django-redis==5.4.0
django-restql==0.15.3
django-simple-captcha==0.5.20
django-timezone-field==6.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
dvadmin-celery==1.0.5
et-xmlfile==1.1.0
fasteners==0.14.1
flower==2.0.1
future==0.18.2
gevent==23.9.1
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httplib2==0.20.2
humanize==4.10.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==24.7.2
inflection==0.5.1
jeepney==0.7.1
keyring==23.5.0
kombu==5.3.7
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.8
mysqlclient==2.1.1
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
openpyxl==3.1.2
packaging==24.1
paramiko==2.9.3
pexpect==4.8.0
Pillow==10.1.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==3.12.4
ptyprocess==0.7.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycups==2.0.1
pyinstaller==6.8.0
pyinstaller-hooks-contrib==2024.7
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==24.2.1
pyparsing==3.1.1
pyPEG2==2.15.2
pypinyin==0.49.0
pyRFC3339==1.1
python-crontab==3.2.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
redis==5.0.8
reportlab==3.6.8
requests==2.31.0
SecretStorage==3.3.1
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.1
tenant-schemas-celery==3.0.0
tomli==2.0.1
tornado==6.4.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2024.1
tzlocal==5.1
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==1.26.5
user-agents==2.2.0
uvicorn==0.23.2
vine==5.1.0
wadllib==1.3.6
wcwidth==0.2.13
websockets==11.0.3
whitenoise==6.6.0
xdg==5
zipp==1.0.0
zope.event==5.0
zope.interface==6.4.post2
